app/requirements.txt
alabaster==0.7.12
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
Babel==2.8.0
bidict==0.22.1
binwalk==2.3.3
blinker==1.6.2
Brlapi==0.8.3
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
configobj==5.0.6
cryptography==3.4.8
cstruct==5.3
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
docutils==0.17.1
email-validator==2.0.0
Flask==2.2.5
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.1.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
htmlmin==0.1.12
httplib2==0.20.2
idna==3.6
imagesize==1.3.0
importlib-metadata==4.6.4
iotop==0.6
itsdangerous==2.1.2
jeepney==0.7.1
jefferson==0.4.1
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lesscpy==0.15.1
louis==3.20.0
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
onboard==1.4.1
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyserial==3.5
pysub==0.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-engineio==4.8.2
python-socketio==5.11.0
python-xapp==2.2.1
pytz==2023.3.post1
pyusb==1.2.1
pyxdg==0.27
PyYAML==5.4.1
rcssmin==1.1.2
referencing==0.32.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
roman==3.3
rpds-py==0.17.1
SecretStorage==3.3.1
setproctitle==1.2.2
simple-websocket==1.0.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.2
SQLAlchemy==2.0.21
terminator==2.1.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==2.1.0
URLObject==2.4.3
wadllib==1.3.6
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
xxhash==3.4.1
zipp==1.0.0
